Rework the CRT showcase: VHS effect, emissive glass tube, environment menu
The screen effect is a VHS-playback shader (tape wave, tape crease,
head-switching noise, chromatic bloom, AC beat), each aspect an
adjustable parameter, running on a lit material: the tube is black
glossy glass and the picture is emitted light, so the environment
glares off the screen. The in-screen panel gains brightness (emissive
scale) and gloss (roughness) sliders alongside the tape controls.

The stress tests' environment switcher (the Khronos sample HDRs, the
resolver with caching, and the HUD popup) moves to a shared example
helper, the stress tests use it, and the widget-texture example gains
the menu so the glare can be tested under different lighting.
